chore(convention) : 패키지 명 변경
- 구글 Java Style Guide에 기준(lowercase)으로 패키지명 변경
  - GlobalFixture to globalfixture
  - reservationTheme to reservationtheme
  - reservationTime to reservationtime
