--- src/main/java/roomescape/reservationTheme/domain/ReservationTheme.java
@@ -1,96 +1,0 @@
-package roomescape.reservationTheme.domain;
-
-import java.util.Objects;
-
-public class ReservationTheme {
-
-    private Long id;
-    private String name;
-    private String description;
-    private String thumbnail;
-
-    public ReservationTheme(String name, String description, String thumbnail) {
-        this.name = name;
-        this.description = description;
-        this.thumbnail = thumbnail;
-    }
-
-    public ReservationTheme(Long id, String name, String description, String thumbnail) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.thumbnail = thumbnail;
-    }
-
-    public static class Builder {
-
-        private Long id;
-        private String name;
-        private String description;
-        private String thumbnail;
-
-        public Builder id(Long id) {
-            this.id = id;
-            return this;
-        }
-
-        public Builder name(String name) {
-            this.name = name;
-            return this;
-        }
-
-        public Builder description(String description) {
-            this.description = description;
-            return this;
-        }
-
-        public Builder thumbnail(String thumbnail) {
-            this.thumbnail = thumbnail;
-            return this;
-        }
-
-        public ReservationTheme build() {
-            return new ReservationTheme(id, name, description, thumbnail);
-        }
-
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getThumbnail() {
-        return thumbnail;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        ReservationTheme that = (ReservationTheme) o;
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-    @Override
-    public String toString() {
-        return "ReservationTheme{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", description='" + description + '\'' +
-                ", thumbnail='" + thumbnail + '\'' +
-                '}';
-    }
-}
