+++ src/main/java/roomescape/reservationtheme/dto/ReservationThemeResponseDto.java
@@ -1,0 +1,42 @@
+package roomescape.reservationtheme.dto;
+
+public class ReservationThemeResponseDto {
+
+    private Long id;
+    private String name;
+    private String description;
+    private String thumbnail;
+
+    public ReservationThemeResponseDto(Long id, String name, String description, String thumbnail) {
+        this.id = id;
+        this.name = name;
+        this.description = description;
+        this.thumbnail = thumbnail;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public String getThumbnail() {
+        return thumbnail;
+    }
+
+    @Override
+    public String toString() {
+        return "{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", description='" + description + '\'' +
+                ", thumbnail='" + thumbnail + '\'' +
+                '}';
+    }
+}
