+++ src/main/java/roomescape/reservationtime/infra/ReservationTimeRepository.java
@@ -1,0 +1,55 @@
+package roomescape.reservationtime.infra;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import roomescape.reservationtime.domain.ReservationTime;
+
+import java.util.List;
+
+@Repository
+public class ReservationTimeRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert simpleJdbcInsert;
+    private final RowMapper<ReservationTime> rowMapper;
+
+    public ReservationTimeRepository(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.rowMapper = (resultSet, rowNum) -> new ReservationTime(
+                resultSet.getLong("id"),
+                resultSet.getString("start_at")
+        );
+        this.simpleJdbcInsert = new SimpleJdbcInsert(jdbcTemplate)
+                .withTableName("reservation_time")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public List<ReservationTime> findAll() {
+        final String sql = "SELECT id, start_at FROM reservation_time";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public Long save(final ReservationTime reservationTime) {
+        final SqlParameterSource parameterSource = new BeanPropertySqlParameterSource(reservationTime);
+        return simpleJdbcInsert.executeAndReturnKey(parameterSource).longValue();
+    }
+
+    public void deleteById(final Long id) {
+        final String sql = "DELETE FROM reservation_time WHERE id = ?";
+        jdbcTemplate.update(sql, Long.valueOf(id));
+    }
+
+    public boolean existsById(final Long id) {
+        final String sql = "SELECT EXISTS(SELECT 1 FROM reservation_time WHERE id = ?)";
+        return jdbcTemplate.queryForObject(sql, Boolean.class, id);
+    }
+
+    public ReservationTime findById(final Long id) {
+        final String sql = "SELECT id, start_at FROM reservation_time WHERE id = ?";
+        return jdbcTemplate.queryForObject(sql, rowMapper, id);
+    }
+}
