--- src/test/java/roomescape/GloblaFixture/dto/ReservationTimeDtoFixture.java
@@ -1,10 +1,0 @@
-package roomescape.GloblaFixture.dto;
-
-import roomescape.reservationTime.dto.ReservationTimeRequestDto;
-
-public class ReservationTimeDtoFixture {
-
-    public static ReservationTimeRequestDto createReservationTimeRequestDto(){
-        return new ReservationTimeRequestDto(1L, "10:00");
-    }
-}
