--- src/test/java/roomescape/GloblaFixture/entity/ReservationTimeFixture.java
@@ -1,10 +1,0 @@
-package roomescape.GloblaFixture.entity;
-
-import roomescape.reservationTime.domain.ReservationTime;
-
-public class ReservationTimeFixture {
-
-    public static ReservationTime createReservationTime(){
-        return new ReservationTime(1L,"10:00");
-    }
-}
