+++ src/test/java/roomescape/globalfixture/entity/ReservationFixture.java
@@ -1,0 +1,15 @@
+package roomescape.globalfixture.entity;
+
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservationtheme.domain.ReservationTheme;
+import roomescape.reservationtime.domain.ReservationTime;
+
+public class ReservationFixture {
+
+    public static Reservation createReservation(){
+        final ReservationTime reservationTime = ReservationTimeFixture.createReservationTime();
+        final ReservationTheme reservationTheme = ReservationThemeFixture.createReservationTheme();
+        return new Reservation("김준성", "2025-08-15", reservationTime, reservationTheme);
+    }
+
+}
