+++ src/test/java/roomescape/globalfixture/entity/ReservationThemeFixture.java
@@ -1,0 +1,10 @@
+package roomescape.globalfixture.entity;
+
+import roomescape.reservationtheme.domain.ReservationTheme;
+
+public class ReservationThemeFixture {
+
+    public static ReservationTheme createReservationTheme(){
+        return new ReservationTheme(1L,"테마1", "설명1", "썸네일1");
+    }
+}
