+++ src/test/java/roomescape/reservation/ReservationControllerTest.java
@@ -8,23 +8,19 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.GloblaFixture.dto.ReservationDtoFixture;
-import roomescape.GloblaFixture.dto.ReservationThemeDtoFixture;
-import roomescape.GloblaFixture.dto.ReservationTimeDtoFixture;
+import roomescape.globalfixture.dto.ReservationDtoFixture;
+import roomescape.globalfixture.dto.ReservationThemeDtoFixture;
+import roomescape.globalfixture.dto.ReservationTimeDtoFixture;
 import roomescape.reservation.dto.ReservationRequestDto;
 import roomescape.reservation.dto.ReservationResponseDto;
-import roomescape.reservation.infra.ReservationRepository;
-import roomescape.reservationTheme.dto.ReservationThemeRequestDto;
-import roomescape.reservationTime.dto.ReservationTimeRequestDto;
-import roomescape.reservationTime.dto.ReservationTimeResponseDto;
+import roomescape.reservationtheme.dto.ReservationThemeRequestDto;
+import roomescape.reservationtime.dto.ReservationTimeRequestDto;
+import roomescape.reservationtime.dto.ReservationTimeResponseDto;
 
 import static org.assertj.core.api.Assertions.assertThat;
